Fixed keybind does not save after loading a new level
Fixed issue of keybind does not save after loading a new level, due to level manager create a new player every time entering a new level. When creating the player it hard code the keybind to the player. Instead now when we create the player it will use pause menu keys for player movements